Prepare for Jquery Joyride onpage help.
I didn't activate the navigation link to start the tour, because I want
to clean up the text inside the tour steps first. But now I can
simply enable the link later when the descriptions are good enough
and my visitors get online help what to do with my page.